esx 1.7.5

+++ esx_multicharacter/server/main.lua
@@ -10,7 +10,7 @@ elseif ESX.GetConfig().Multichar == true then
 			DATABASE = connectionString:sub(connectionString:find('/')+1, -1):gsub('[%?]+[%w%p]*$', '')
 		else
 			connectionString = {string.strsplit(';', connectionString)}
-			for i=1, #connectionString do
+			for i = 1, #connectionString do
 				local v = connectionString[i]
 				if v:match('database') then
 					DATABASE = v:sub(10, #v)
@@ -19,7 +19,7 @@ elseif ESX.GetConfig().Multichar == true then
 		end
 	end
 
-	local DB_TABLES = {}
+	local DB_TABLES = {users = 'identifier'}
 	local FETCH = nil
 	local SLOTS = Config.Slots or 4
 	local PREFIX = Config.Prefix or 'char'
@@ -48,48 +48,51 @@ elseif ESX.GetConfig().Multichar == true then
 		local identifier = GetIdentifier(source)
 		ESX.Players[identifier] = true
 
-		local slots = MySQL.Sync.fetchScalar("SELECT slots FROM multicharacter_slots WHERE identifier = ?", {
-			identifier
-		}) or SLOTS
+		local slots = MySQL.scalar.await('SELECT slots FROM multicharacter_slots WHERE identifier = ?', { identifier }) or SLOTS
 		identifier = PREFIX..'%:'..identifier
 
-		MySQL.Async.fetchAll(FETCH, {identifier, slots}, function(result)
-			local characters
-			if result then
-				local characterCount = #result
-				characters = table.create(0, characterCount)
-				for i=1, characterCount, 1 do
-					local i = result[i]
-					local job, grade = i.job or 'unemployed', tostring(i.job_grade)
-					if ESX.Jobs[job] and ESX.Jobs[job].grades[grade] then
-						if job ~= 'unemployed' then grade = ESX.Jobs[job].grades[grade].label else grade = '' end
-						job = ESX.Jobs[job].label
-					end
-					local accounts = json.decode(i.accounts)
-					local id = tonumber(string.sub(i.identifier, #PREFIX+1, string.find(i.identifier, ':')-1))
-					characters[id] = {
-						id = id,
-						bank = accounts.bank,
-						money = accounts.money,
-						job = job,
-						job_grade = grade,
-						firstname = i.firstname,
-						lastname = i.lastname,
-						dateofbirth = i.dateofbirth,
-						skin = json.decode(i.skin),
-						disabled = i.disabled,
-						sex = i.sex == 'm' and _('male') or _('female')
-					}
+		local result = MySQL.query.await(FETCH, {identifier, slots})
+		local characters
+
+		if result then
+			local characterCount = #result
+			characters = table.create(0, characterCount)
+
+			for i = 1, characterCount, 1 do
+				local v = result[i]
+				local job, grade = v.job or 'unemployed', tostring(v.job_grade)
+
+				if ESX.Jobs[job] and ESX.Jobs[job].grades[grade] then
+					if job ~= 'unemployed' then grade = ESX.Jobs[job].grades[grade].label else grade = '' end
+					job = ESX.Jobs[job].label
 				end
+
+				local accounts = json.decode(v.accounts)
+				local id = tonumber(string.sub(v.identifier, #PREFIX+1, string.find(v.identifier, ':')-1))
+
+				characters[id] = {
+					id = id,
+					bank = accounts.bank,
+					money = accounts.money,
+					job = job,
+					job_grade = grade,
+					firstname = v.firstname,
+					lastname = v.lastname,
+					dateofbirth = v.dateofbirth,
+					skin = v.skin and json.decode(v.skin) or {},
+					disabled = v.disabled,
+					sex = v.sex == 'm' and _('male') or _('female')
+				}
 			end
-			TriggerClientEvent('esx_multicharacter:SetupUI', source, characters, slots)
-		end)
+		end
+
+		TriggerClientEvent('esx_multicharacter:SetupUI', source, characters, slots)
 	end
 
 	AddEventHandler('playerConnecting', function(playerName, setKickReason, deferrals)
 		deferrals.defer()
 		local identifier = GetIdentifier(source)
-		Citizen.Wait(100)
+
 		if identifier then
 			if ESX.Players[identifier] then
 				deferrals.done(('A player is already connected to the server with this identifier.\nYour identifier: %s:%s'):format(PRIMARY_IDENTIFIER, identifier))
@@ -103,16 +106,16 @@ elseif ESX.GetConfig().Multichar == true then
 
 	local function DeleteCharacter(source, charid)
 		local identifier = ('%s%s:%s'):format(PREFIX, charid, GetIdentifier(source))
-		local query = "DELETE FROM ?? WHERE ?? = ?"
-		local tableCount = #DB_TABLES
-		local queries = table.create(tableCount, 0)
+		local query = 'DELETE FROM %s WHERE %s = ?'
+		local queries = {}
+		local count = 0
 
-		for i=1, tableCount do
-			local v = DB_TABLES[i]
-			queries[i] = {query = query, values = {v.table, v.column, identifier}}
+		for table, column in pairs(DB_TABLES) do
+			count += 1
+			queries[count] = {query = query:format(table, column), values = {identifier}}
 		end
 
-		MySQL.Async.transaction(queries, function(result)
+		MySQL.transaction(queries, function(result)
 			if result then
 				print(('[^2INFO^7] Player [%s] %s has deleted a character (%s)'):format(GetPlayerName(source), source, identifier))
 				Citizen.Wait(50)
@@ -124,53 +127,64 @@ elseif ESX.GetConfig().Multichar == true then
 	end
 
 	MySQL.ready(function()
-		MySQL.Async.fetchAll('SELECT TABLE_NAME, COLUMN_NAME, CHARACTER_MAXIMUM_LENGTH FROM INFORMATION_SCHEMA.COLUMNS WHERE TABLE_SCHEMA = ? AND (COLUMN_NAME = ? OR COLUMN_NAME = ?) ', {
-			DATABASE, 'identifier', 'owner'
-		}, function(result)
-			if result then
-				local varchar, varsize = {}, 0
-				for k, v in pairs(result) do
-					if v.CHARACTER_MAXIMUM_LENGTH and v.CHARACTER_MAXIMUM_LENGTH >= 40 and v.CHARACTER_MAXIMUM_LENGTH < 60 then varchar[v.TABLE_NAME] = v.COLUMN_NAME varsize = varsize+1 end
-					table.insert(DB_TABLES, {table = v.TABLE_NAME, column = v.COLUMN_NAME})
-				end
-				if next(varchar) then
-					local query = "ALTER TABLE ?? MODIFY COLUMN ?? VARCHAR(60)"
-					local queries = table.create(varsize, 0)
+		local length = 42 + #PREFIX
+		local DB_COLUMNS = MySQL.query.await(('SELECT TABLE_NAME, COLUMN_NAME, CHARACTER_MAXIMUM_LENGTH FROM INFORMATION_SCHEMA.COLUMNS WHERE TABLE_SCHEMA = "%s" AND DATA_TYPE = "varchar" AND COLUMN_NAME IN (?)'):format(DATABASE, length), {
+			{'identifier', 'owner'}
+		})
 
-					for k, v in pairs(varchar) do
-						queries[#queries+1] = {query = query, values = {k, v}}
-					end
+		if DB_COLUMNS then
+			local columns = {}
+			local count = 0
 
-					MySQL.Async.transaction(queries, function(result)
-						if result then
-							print(('[^2INFO^7] Updated ^5%s^7 columns to use VARCHAR(60)'):format(varsize))
-						else
-							print(('[^2INFO^7] Unable to update ^5%s^7 columns to use VARCHAR(60)'):format(varsize))
-						end
-					end)
+			for i = 1, #DB_COLUMNS do
+				local v = DB_COLUMNS[i]
+
+				if v?.CHARACTER_MAXIMUM_LENGTH ~= length then
+					count += 1
+					columns[v.TABLE_NAME] = v.COLUMN_NAME
 				end
-				ESX.Jobs = {}
-				local function GetJobs()
-					if ESX.GetJobs then return ESX.GetJobs()
-					else return exports['es_extended']:getSharedObject().Jobs end
-				end
-				repeat
-					ESX.Jobs = GetJobs()
-					Citizen.Wait(50)
-				until next(ESX.Jobs)
-				FETCH = MySQL.Sync.store("SELECT identifier, accounts, job, job_grade, firstname, lastname, dateofbirth, sex, skin, disabled FROM users WHERE identifier LIKE ? LIMIT ?")
 			end
-		end)
+
+			if next(columns) then
+				local query = 'ALTER TABLE `%s` MODIFY COLUMN `%s` VARCHAR(%s)'
+				local queries = table.create(count, 0)
+
+				for k, v in pairs(columns) do
+					queries[#queries+1] = {query = query:format(k, v, length)}
+				end
+
+				if MySQL.transaction.await(queries) then
+					print(('[^2INFO^7] Updated ^5%s^7 columns to use VARCHAR(%s)'):format(count, length))
+				else
+					print(('[^2INFO^7] Unable to update ^5%s^7 columns to use VARCHAR(%s)'):format(count, length))
+				end
+			end
+
+			ESX.Jobs = {}
+
+			local function GetJobs()
+				if ESX.GetJobs then
+					return ESX.GetJobs()
+				end
+				return exports.es_extended:getSharedObject().Jobs
+			end
+
+			repeat
+				ESX.Jobs = GetJobs()
+				Citizen.Wait(50)
+			until next(ESX.Jobs)
+
+			FETCH = 'SELECT identifier, accounts, job, job_grade, firstname, lastname, dateofbirth, sex, skin, disabled FROM users WHERE identifier LIKE ? LIMIT ?'
+		end
 	end)
 
-	RegisterServerEvent("esx_multicharacter:SetupCharacters")
-	AddEventHandler('esx_multicharacter:SetupCharacters', function()
+	RegisterNetEvent('esx_multicharacter:SetupCharacters', function()
 		SetupCharacters(source)
 	end)
 
 	local awaitingRegistration = {}
-	RegisterServerEvent("esx_multicharacter:CharacterChosen")
-	AddEventHandler('esx_multicharacter:CharacterChosen', function(charid, isNew)
+
+	RegisterNetEvent('esx_multicharacter:CharacterChosen', function(charid, isNew)
 		if type(charid) == 'number' and string.len(charid) <= 2 and type(isNew) == 'boolean' then
 			if isNew then
 				awaitingRegistration[source] = charid
@@ -192,18 +206,16 @@ elseif ESX.GetConfig().Multichar == true then
 		ESX.Players[GetIdentifier(source)] = nil
 	end)
 
-	RegisterServerEvent("esx_multicharacter:DeleteCharacter")
-	AddEventHandler('esx_multicharacter:DeleteCharacter', function(charid)
-		if type(charid) == "number" and string.len(charid) <= 2 then
+	RegisterNetEvent('esx_multicharacter:DeleteCharacter', function(charid)
+		if Config.CanDelete and type(charid) == 'number' and string.len(charid) <= 2 then
 			DeleteCharacter(source, charid)
 		end
 	end)
 
-	RegisterServerEvent("esx_multicharacter:relog")
-	AddEventHandler('esx_multicharacter:relog', function()
+	RegisterNetEvent('esx_multicharacter:relog', function()
 		TriggerEvent('esx:playerLogout', source)
 	end)
 
-else
-	assert(nil, '^3WARNING: Multicharacter is disabled - please check your ESX configuration^0')
-end
+	else
+		assert(nil, '^3WARNING: Multicharacter is disabled - please check your ESX configuration^0')
+	end
